Conflicts:
	GitCommands/GitCommands.sln

+++ GitUI/Gravatar.cs
@@ -198,7 +198,11 @@ namespace GitUI
         {
             string imageFileName = string.Concat(theEmail, ".png");
             IsolatedStorageFile isolatedStorage = IsolatedStorageFile.GetStore(IsolatedStorageScope.User | IsolatedStorageScope.Assembly, null, null);
-            isolatedStorage.DeleteFile(imageFileName);
+            
+            if (isolatedStorage.GetFileNames(imageFileName).Length != 0)
+            {
+                isolatedStorage.DeleteFile(imageFileName);
+            }
 
             UpdateGravatar();
         }
@@ -231,6 +235,15 @@ namespace GitUI
             IsolatedStorageFile isolatedStorage = IsolatedStorageFile.GetStore(IsolatedStorageScope.User | IsolatedStorageScope.Assembly, null, null);
             foreach (string gravatarFileName in isolatedStorage.GetFileNames("*.png"))
                 isolatedStorage.DeleteFile(gravatarFileName);
+        }
+
+        private void smallToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            ToolStripItem toolStripItem = (ToolStripItem)sender;
+
+            GitCommands.Settings.AuthorImageSize = int.Parse(toolStripItem.Text);
+            ClearImageCache();
+            UpdateGravatar();
         }
     }
 }
